Refactoring of UI.java
An attempt to modularize UI.java to limit navigation issues when adding new features.

JFrame structure has been changed, previous JPanel cards (cardLayout) has been split into 2 JFrames; EntryPointFrame and DashboardFrame.

Issue:
When switching between Login and Register cards, the common components (on the JFrame) do not appropriately paint -attributed to mismanagement of the ActionPerformed.
